Refactor OpalDefendantAccountHistoryService to integrate DefendantTransactionHistorySourceService for fetching transactions
